Added correctFilterTest

+++ src/test/java/page/AbstractPage.java
@@ -25,7 +25,7 @@ public abstract class AbstractPage {
         	.until(ExpectedConditions.elementToBeClickable(By.xpath(locator)));
     }
 
-    protected WebElement waitForElementVisibility(WebElement element, int time) {
+    protected WebElement waitForElementVisibility(WebElement element, long time) {
         return new WebDriverWait(driver, Duration.ofSeconds(time))
                 .until(ExpectedConditions.visibilityOf(element));
     }
